requirements.txt
streamlit
requests
Pillow
deep-translator